refactor: 모델 디렉토리 구조 변경 및 이름 수정
