--- WeatherForecast/WeatherForecast/Model/Coordinate.swift
@@ -1,18 +1,0 @@
-//
-//  Coordinate.swift
-//  WeatherForecast
-//
-//  Created by 샬롯, 수박, 루얀 on 2021/09/28.
-//
-
-import Foundation
-
-struct Coordinate: Codable {
-    let longitude: Double
-    let latitude: Double
-    
-    enum CodingKeys: String, CodingKey {
-        case longitude = "lon"
-        case latitude = "lat"
-    }
-}
